fix: create file without jsonfilr format, deleted CreateFile and FileUploader file and made createFileMethod for both duplicate and create file
