+++ src/components/file/index.js
@@ -1,7 +1,5 @@
 export { default as DeleteFile } from "./Delete.vue";
 export { default as ActionTabFileEditor } from "./ActionTabFileEditor.vue";
-export { default as CreateFile } from "./CreateFile.vue";
-export { default as FileUploader } from "./FileUploader.vue";
 export { default as DuplicateFile } from "./DuplicateFile.vue";
 export { default as RenameFile } from "./Rename.vue";
 export { default as PreviewBtn } from "./Preview.vue";
